requirements.txt
fastapi
uvicorn[standard]
openai
supabase
python-dotenv
tqdm
httpx
requests
numpy
tiktoken